Extract from existing sources
I tried this to be a grafted history because MUST NOT include all the other,
unrelated crap from the mess where I did extract this code from. But it seems a
hard thing to do as all sorts of git commands (eg push, clone, ...) did not
work as epxected. So I keep the commit metadata here where this got split-off.

  commit 196ff8beda7a23aee6e1e7388cfc24f6247a63c9
  Author: Andreas Fankhauser (@tux-six) <https://hiddenalpha.ch/slnk/id/1:c0n74ct>
  Date:   2022-06-10 02:52:11 +0200

      (scriptlee) Fix bug in http